Supply a strategy for obscuring data per obscured field
Allow passing in either a list of fields or a map of fields to strategies. Default to the previously existing strategy of shuffling the characters in a field. Being able to pass in a strategy is useful in case we are moving data including personally identifiable information to a less-secure environment, such as moving user names or emails to a testing environment for further experimentation.